Float Printing Support (#6)
Add native library, currently has a mini version of the dotnet MathHelpers.cpp which contains a single function that isn't included otherwise despite other functions in the same file appearing to work.

Updated the build script to build the native library and link it to the rest of the project.