Added remaining functions to Random
That is with the exception of the version of NextBytes that uses a span since that is not supported current.

The only function that might be a bit weird is NextInt64(long, long) since the range can not be stored in an signed format and even with the work around, the result could overflow long. Can it overflow double?