Updated Console.ReadKey
ReadKey should match dotnet behaviour for space and tab now now. Additionally the cursor starts enabled by default now.

Currently the other console reading functions are disabled because the internal char buffer has been removed. The plan is to see if the console can function without it and if not add a proper circular deque.