Cleaned up console output
Added list of supported drawing chars from unicode, more are probably supported but this is the minimum allowed by spec. Not sure if this needs to be added but it is nice to have a list.

Increased depth of function documentation. Not sure if i went overboard.

Moved console output structs to new folder, currently this does not add a namespace but I might change that later on.