Enabled Cursor and Lock Key Checks
Number and Caps lock checks have been made somewhat more consistent by noticing that ReadKeyStrokeEx can still return valid data even if not returning EFI_SUCCESS. There are still some issues as noted in the todos but I expect that the interrupt approach can deal with this.

The CursorSize field has been hardcoded to what appears to be a reasonable value since it can not be changed at runtime currently. If I end up mimicking dotnet behaviour and using a timer event to blink the cursor, it might be possible to use specific block characters to fake having a customisable cursor. However this depends on character support on specific output devices.