Checking in work on efi buffer replacement
While Console.ReadKey is easy to implement without a buffer, it have proven largely impossible to implement Console.Read this way. This is because while it is possible to wait for enter being pressed after the function begins and then returning the first char returned from efi, that ignores the fact that enter could have been pressed before the function began and that this function is meant to print all chars between the first and enter while still leaving them for a latter call. This is impossible with the efi buffer and so i am taking the annoying approach of interrupting on all possible inputs and completely ignoring the efi buffer.

Currently this just shows that the buffer and interrupts can be enabled and then chars can be read in continually.

Remnants of previous attempts litter this version of Console.cs and Class1.cs and will be stripped out during a later commit one I get this approach working correctly.