Fixed backspace printing bug in Read
Previously pressing backspace before any other char could cause Read to return 0 instead of any chars pressed afterward.

Partially fixed KeyAvailable. Depending on what is needed for ReadKey, this might be updated to account for the buffer as well. It is always going to be sketchy in that sometimes it will miss chars due to limitations of the efi console in protocol.