Fixed removal of tabs in Read
Previously each backspace removed one space from tab which is different from the dotnet implementation which removes the entire tab in one go.
This required storing the number of spaces the tab added which I placed in a new struct with the char it self, this ended up being smaller than the previous EFI_KEY_DATA struct and so the stack usage of the char buffer is 1/3 of it previous value.