Readded NumberLock and CapsLock bools
Unlike the previous methods that could remove a key from the efi buffer, this method results in leaves the buffer intact and gives an almost live view of the lock key's status. This method however is only useful of EFI_KEY_STATE_EXPOSED works on the target system. This is probably fine since if not the only option is destructive.

 It might be possible to reduce any performance impact by checking for all 4 possible options directly and not matching all incomplete keys which includes shift, alt, ctrl and scroll lock.