Added rough method of automatic debugger starting
Ideally the main visual studio instance would start debugging but I can't figure out how.
Best I can get for now is opening a new instance and getting it to start debugging with a script.
This currently requires git grep but there are a couple of msbuild extensions with similar functionality on nuget that could be used instead at some point.
Tried to ensure the new vs instance was brought to front but appears quite inconsistent.